requirements.txt
mysql.connector
requests
json
tabulate
